Begin adding core calculator logic to the backend of the application.
Added classes to deal with operations that deal with two, three, or n number of operants. Need to add Unary class for some functions (e.g. square root).
NotANumberException added to project to specify instances of ArithmeticException where NaN is produced. NaN can be a valid result for a calculator, and should not be dismissed or removed.